+++ titiler/endpoints/factory.py
@@ -36,12 +36,7 @@ from ..dependencies import (
 )
 from ..models.mapbox import TileJSON
 from ..models.OGC import TileMatrixSetList
-from ..resources.enums import (
-    ImageType,
-    MimeTypes,
-    OptionalHeaders,
-    PixelSelectionMethod,
-)
+from ..resources.enums import ImageType, MimeType, OptionalHeader, PixelSelectionMethod
 from ..resources.responses import GeoJSONResponse, XMLResponse
 from ..templates import templates
 
@@ -106,7 +101,7 @@ class BaseTilerFactory(metaclass=abc.ABCMeta):
     gdal_config: Dict = field(default_factory=dict)
 
     # add additional headers in response
-    optional_headers: List[OptionalHeaders] = field(default_factory=list)
+    optional_headers: List[OptionalHeader] = field(default_factory=list)
 
     def __post_init__(self):
         """Post Init: register route and configure specific options."""
@@ -349,7 +344,7 @@ class TilerFactory(BaseTilerFactory):
                 )
             timings.append(("format", round(t.elapsed * 1000, 2)))
 
-            if OptionalHeaders.server_timing in self.optional_headers:
+            if OptionalHeader.server_timing in self.optional_headers:
                 headers["Server-Timing"] = ", ".join(
                     [f"{name};dur={time}" for (name, time) in timings]
                 )
@@ -516,7 +511,7 @@ class TilerFactory(BaseTilerFactory):
                     "layer_name": "cogeo",
                     "media_type": tile_format.mimetype,
                 },
-                media_type=MimeTypes.xml.value,
+                media_type=MimeType.xml.value,
             )
 
     ############################################################################
@@ -553,7 +548,7 @@ class TilerFactory(BaseTilerFactory):
                         )
             timings.append(("dataread", round(t.elapsed * 1000, 2)))
 
-            if OptionalHeaders.server_timing in self.optional_headers:
+            if OptionalHeader.server_timing in self.optional_headers:
                 response.headers["Server-Timing"] = ", ".join(
                     [f"{name};dur={time}" for (name, time) in timings]
                 )
@@ -616,7 +611,7 @@ class TilerFactory(BaseTilerFactory):
                 )
             timings.append(("format", round(t.elapsed * 1000, 2)))
 
-            if OptionalHeaders.server_timing in self.optional_headers:
+            if OptionalHeader.server_timing in self.optional_headers:
                 headers["Server-Timing"] = ", ".join(
                     [f"{name};dur={time}" for (name, time) in timings]
                 )
@@ -684,7 +679,7 @@ class TilerFactory(BaseTilerFactory):
                 )
             timings.append(("format", round(t.elapsed * 1000, 2)))
 
-            if OptionalHeaders.server_timing in self.optional_headers:
+            if OptionalHeader.server_timing in self.optional_headers:
                 headers["Server-Timing"] = ", ".join(
                     [f"{name};dur={time}" for (name, time) in timings]
                 )
@@ -1140,12 +1135,12 @@ class MosaicTilerFactory(BaseTilerFactory):
                 )
             timings.append(("format", round(t.elapsed * 1000, 2)))
 
-            if OptionalHeaders.server_timing in self.optional_headers:
+            if OptionalHeader.server_timing in self.optional_headers:
                 headers["Server-Timing"] = ", ".join(
                     [f"{name};dur={time}" for (name, time) in timings]
                 )
 
-            if OptionalHeaders.x_assets in self.optional_headers:
+            if OptionalHeader.x_assets in self.optional_headers:
                 headers["X-Assets"] = ",".join(data.assets)
 
             return Response(content, media_type=format.mimetype, headers=headers)
@@ -1308,7 +1303,7 @@ class MosaicTilerFactory(BaseTilerFactory):
                     "layer_name": "cogeo",
                     "media_type": tile_format.mimetype,
                 },
-                media_type=MimeTypes.xml.value,
+                media_type=MimeType.xml.value,
             )
 
     ############################################################################
@@ -1354,7 +1349,7 @@ class MosaicTilerFactory(BaseTilerFactory):
                         )
             timings.append(("dataread", round((t.elapsed - mosaic_read) * 1000, 2)))
 
-            if OptionalHeaders.server_timing in self.optional_headers:
+            if OptionalHeader.server_timing in self.optional_headers:
                 response.headers["Server-Timing"] = ", ".join(
                     [f"{name};dur={time}" for (name, time) in timings]
                 )